# swagger.YAML
swagger: "2.0"
info:
  description: "Bu API scraper mantığı ile android ve ios mağaza yorumlarını ve uygulama bilgilerini döndürür."
  version: "1.1.1"
  title: "Reviews API"
  termsOfService: "https://reviewsapiprod.herokuapp.com"
  contact:
    email: "https://reviewsapiprod.herokuapp.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "reviewsapiprod.herokuapp.com"
basePath: "/api/v1"
tags:
  - name: "androidreviewlist"
    description: "Operations about user"
    externalDocs:
      description: "Find out more about our store"
      url: "http://swagger.io"
schemes:
  - "https"
  - "http"
paths:
  /androidreviewlist/limit={limit}/sorting={sorting}/nextPaginationToken={nextPaginationToken}/packageName={packageName}:
    get:
      tags:
        - "androidreviewlist"
      summary: "Get all reviews"
      description: ""
      operationId: "androidreviewlist"
      produces:
        - "application/json"
      parameters:
        - name: "packageName"
          in: "path"
          description: "The name that needs to be fetched. Use user1 for testing. "
          required: true
          type: "string"
        - name: "limit"
          in: "path"
          description: "The name that needs to be fetched. Use user1 for testing. "
          required: true
          type: "string"
        - name: "sorting"
          in: "path"
          description: "The name that needs to be fetched. Use user1 for testing. "
          required: true
          type: "string"
        - name: "nextPaginationToken"
          in: "path"
          description: "The name that needs to be fetched. Use user1 for testing. "
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/data"
        "400":
          description: "Invalid data supplied"
        "404":
          description: "data not found"
definitions:
  data:
    type: "array"
    items:
      properties:
        id:
          type: "string"
        userName:
          type: "string"
        userImage:
          type: "string"
        date:
          type: "string"
        score:
          type: "integer"
          format: "int32"
        scoreText:
          type: "string"
        url:
          type: "string"
        title:
          type: "string"
        text:
          type: "string"
        replyDate:
          type: "string"
        replyText:
          type: "string"
        version:
          type: "string"
        thumbsUp:
          type: "string"
        criterias:
          type: "string"
  nextPaginationToken:
    type: "string"